create login and signOut methods
